Fix Deployment manifest to remove erroneous variable substitutions

+++ sync/patches/kustomize/patch.sh
@@ -18,6 +18,17 @@ cp "${VENDIR_DIR}"/control-plane-components.yaml "${script_dir}"/tmp/input/
 
 kustomize build "${script_dir}" -o "${script_dir}"/tmp/output/
 
+# process deployment file to remove bash variable substitutions
+DEPLOYMENT_FILE=$(find "${script_dir}"/tmp/output/ -iname \*deployment\*)
+
+if [ -z "${DEPLOYMENT_FILE}" ]; then
+    echo "Could not find deployment file in the kustomize output."
+    exit 1
+fi
+
+sed -i -E 's/\$\{[^:]+:=([^}]+)\}/\1/g' "${DEPLOYMENT_FILE}"
+# finished processing deployment file
+
 cp "${script_dir}"/tmp/output/*.yaml "${CHART_DIR}"/templates/
 
 rm -rf "${script_dir}"/tmp/
